Write MSGraphAPI class and refactor
